refactor: rename app module to cli
