+++ src/lib.rs
@@ -1,7 +1,7 @@
 extern crate core;
 
-pub mod app;
 pub mod args_format;
+pub mod cli;
 pub mod config_files;
 mod defaults;
 pub mod tasks;
